Run Prettier for Contracts and Scripts.

+++ contracts/interfaces/IBentureAdmin.sol
@@ -2,11 +2,8 @@
 
 pragma solidity ^0.8.0;
 
-
 /// @title An interface of a factory of custom ERC20 tokens
 interface IBentureAdmin {
-
-
     /// @notice Checks it the provided address owns any admin token
     function checkOwner(address user) external view;
 
@@ -15,10 +12,13 @@ interface IBentureAdmin {
     /// @param ERC20Address The address of the potentially controlled ERC20 token
     function verifyAdminToken(address user, address ERC20Address) external view;
 
-    /// @notice Returns the address of the controlled ERC20 token 
+    /// @notice Returns the address of the controlled ERC20 token
     /// @param tokenId The ID of ERC721 token to check
     /// @return The address of the controlled ERC20 token
-    function getControlledAddressById(uint256 tokenId) external view returns (address);
+    function getControlledAddressById(uint256 tokenId)
+        external
+        view
+        returns (address);
 
     /// @notice Mints a new ERC721 token with the address of the controlled ERC20 token
     /// @param to The address of the receiver of the token
@@ -28,23 +28,20 @@ interface IBentureAdmin {
     /// @notice Burns the token with the provided ID
     /// @param tokenId The ID of the token to burn
     function burn(uint256 tokenId) external;
-    
+
     /// @dev Indicates that a new ERC721 token got minted
     event AdminTokenCreated(
         uint256 indexed tokenId,
         address indexed ERC20Address
     );
-    
+
     /// @dev Indicates that an ERC721 token got burnt
-    event AdminTokenBurnt(
-        uint256 indexed tokenId
-    );
+    event AdminTokenBurnt(uint256 indexed tokenId);
 
     /// @dev Indicates that an ERC721 token got transferred
     event AdminTokenTransfered(
-        address indexed from, 
-        address indexed to, 
-        uint tokenId
+        address indexed from,
+        address indexed to,
+        uint256 tokenId
     );
-
 }
